Fix lms js i18n static collection.

+++ pavelib/assets.py
@@ -679,11 +679,17 @@ def collect_assets(systems, settings, **kwargs):
             current_sys=sys
         ))
     for sys in systems:
-        os.system("rsync -av {static_collector_dir}/{current_sys}/* {platform_static_dir}/{current_sys}/ --delete-after".format(
-            static_collector_dir=django_settings.STATIC_ROOT_BASE,
-            platform_static_dir=django_settings.EDX_PLATFORM_STATIC_ROOT_BASE,
-            current_sys=sys
-        ))
+        if (sys == 'lms'):
+            os.system("rsync -av {static_collector_dir}/js/i18n/* {platform_static_dir}/cms/js/i18n/ --delete-after".format(
+                static_collector_dir=django_settings.STATIC_ROOT_BASE,
+                platform_static_dir=django_settings.EDX_PLATFORM_STATIC_ROOT_BASE,
+                current_sys=sys
+            ))
+            os.system("rsync -av {static_collector_dir}/{current_sys}/* {platform_static_dir}/{current_sys}/ --delete-after".format(
+                static_collector_dir=django_settings.STATIC_ROOT_BASE,
+                platform_static_dir=django_settings.EDX_PLATFORM_STATIC_ROOT_BASE,
+                current_sys=sys
+            ))
 
 def _collect_assets_cmd(system, **kwargs):
     """
